backend/djangoapp/requirements.txt
Django>=4.2.1,<4.3
psycopg2-binary>=2.9.6,<2.10
djangorestframework>=3.14.0,<3.15
django-cors-headers>=4.0.0,<4.1
djangorestframework-simplejwt>=5.2.2,<5.3
python-decouple==3.8
